refactor: move TransactionSubmittedContent to orderProgressBar module
