+++ apps/cowswap-frontend/src/modules/account/index.ts
@@ -2,3 +2,5 @@ export { OrdersPanel } from './containers/OrdersPanel'
 export { useToggleAccountModal } from './hooks/useToggleAccountModal'
 export { useAccountModalState } from './hooks/useAccountModalState'
 export { AccountIcon } from './containers/AccountDetails/AccountIcon'
+export { GnosisSafeTxDetails } from './containers/Transaction/ActivityDetails'
+export { TransactionInnerDetail } from './containers/Transaction/styled'
