+++ apps/cowswap-frontend/src/modules/orderProgressBar/index.ts
@@ -1,1 +1,5 @@
+// Containers
 export { SurplusModalSetup } from './containers/SurplusModalSetup'
+
+// Hooks
+export { useOrderSubmittedContent } from './hooks/useOrderSubmittedContent'
