+++ apps/cowswap-frontend/src/modules/orderProgressBar/pure/TransactionSubmittedContent/index.cosmos.tsx
