+++ apps/cowswap-frontend/src/modules/trade/index.ts
@@ -30,7 +30,6 @@ export * from './hooks/useBuildTradeDerivedState'
 export * from './hooks/useOnCurrencySelection'
 export * from './hooks/useDerivedTradeState'
 export * from './hooks/useNavigateToNewOrderCallback'
-export * from './hooks/useOrderSubmittedContent'
 export * from './hooks/useIsEoaEthFlow'
 export * from './hooks/useShouldPayGas'
 export * from './hooks/useWrappedToken'
