+++ apps/cowswap-frontend/src/modules/wallet/index.ts
@@ -1,0 +1,1 @@
+export { WatchAssetInWallet } from './containers/WatchAssetInWallet'
